+++ plone/recipe/codeanalysis/analyser.py
@@ -215,11 +215,10 @@ class Analyser:
         """Run the analyser command using options.
 
         Run the analyser command using options and return if the result was
-        sucessful. The analyser command is defined by the 'bin' option, it can
-        be any one of the external analyser commands, as csslint, jshint,
-        flake8, for example. The return code is stored on the return_code
-        attribute. The output is returned but is also stored in the output
-        attribute.
+        successful. The analyser command is defined by the 'bin' option, it can
+        be any one of the external analyser commands, flake8, for example.
+        The return code is stored on the return_code attribute. The output is
+        returned but is also stored in the output attribute.
 
         :return: It return the output of the analyser command.
         """